webapp/requirement.txt
django
django-email-confirmation
django-timezones
pytz
django-social-auth